refactor: 빌드 경고 개선 등
- `next.config.ts` 설정 추가
  - 다음의 경고가 발생하지 않도록 개선합니다.
  - `<w> [webpack.cache.PackFileCacheStrategy] Serializing big strings (143kiB) impacts deserialization performance (consider using Buffer instead and decode when needed)`
- repository root 에 놓인 `mdx-components.js`를 `src/mdx-components.js`로 이동합니다.
  - root 에 놓이지 않아도 되는 코드입니다.
